Rework rules for Belarusian and update phonemes
This patch strip out dependency on Russian, as its complicated relationship
between spelling and pronunciation overburden rules for Belarusian. The rules
themselves are also added. They don't have devoicing at the end of the words
for now. Also, gemination of consonants doesn't work as expected, as it seems
on the top of this rule some other rules apply. This after-processing is
responsible also for the strange behaviour of ending [a].